refactor: reorganize billing tests location
Moves billing-related tests into a dedicated package to improve
project structure. Introduces a new test for billing service
to verify stability under normal and null conditions.
